refactor: 階層構造を整理
