--- data/src/main/java/com/nemo/data/apis/Endpoints.kt
@@ -1,5 +1,0 @@
-package com.nemo.data.apis
-
-object Endpoints {
-    const val githubApiEndpoint = "https://api.github.com/"
-}
