+++ data/src/main/java/com/nemo/data/apis/RetrofitManager.kt
@@ -1,9 +1,10 @@
 package com.nemo.data.apis
 
-import com.nemo.data.apis.Endpoints.githubApiEndpoint
+import com.nemo.data.apis.interfaces.GithubApi
 import retrofit2.Retrofit
 
 object RetrofitManager {
+    private const val githubApiEndpoint = "https://api.github.com/"
     private val retrofit = Retrofit.Builder()
         .baseUrl(githubApiEndpoint)
         .build()
