+++ data/src/main/java/com/nemo/data/apis/interfaces/GithubApi.kt
@@ -1,5 +1,6 @@
-package com.nemo.data.apis
+package com.nemo.data.apis.interfaces
 
+import com.nemo.data.apis.responses.FetchAllProjectsResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
